Made an abstract class for audio
- player would inherit from it
- so will weapon
