+++ Assets/Scripts/AudioPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6ba8ef9baa92b947aed5d782b1d5c14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
