+++ Assets/Scripts/Weapon/NormalBullet.cs
@@ -31,4 +31,18 @@ public class NormalBullet : Bullet
             rgdbdy2d.MovePosition(transform.position + BulletData.BulletSpeed * transform.right * Time.fixedDeltaTime);
         }
     }
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if(collision.gameObject.layer == LayerMask.NameToLayer("Obstacles"))
+        {
+            HitObstacle();
+        }
+        Destroy(gameObject);
+    }
+
+    private void HitObstacle()
+    {
+        Debug.Log("Obstacle Hit");
+    }
 }
